make the feature of LoggerMeili work well
